refactor: 불필요한 상수 파일 분리 하나로 합치기
