--- src/constants/filteringCategories.ts
@@ -1,5 +1,0 @@
-import CATEGORIES from './categories';
-
-const FILTERING_CATEGORIES = ['전체', ...CATEGORIES] as const;
-
-export default FILTERING_CATEGORIES;
