--- src/constants/filteringOrders.ts
@@ -1,3 +1,0 @@
-const FILTERING_ORDERS = ['이름순', '거리순'] as const;
-
-export default FILTERING_ORDERS;
